requirements.txt
Flask==1.1.2
hydra-core==1.0.6
jupyter==1.0.0
matplotlib==3.3.4
numpy==1.19.2
omegaconf==2.0.6
pandas==1.2.1
PyYAML==5.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
sentencepiece==0.1.95
tensorboard==2.5.0
torch==1.6.0
tqdm==4.56.0
transformers==4.3.2
urllib3==1.26.3
gitpython==3.1.18
tensorflow-gpu==2.5.1
grpcio==1.34.0
subword-nmt==0.3.7
